requirements.txt
streamlit>=1.28.0
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
pandas>=1.3.0
transformers>=4.21.0
torch>=1.12.0